requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
google-generativeai
python-dotenv